+++ src/main/java/hr/avrbanac/inanis/internals/Engine.java
@@ -1,6 +1,10 @@
 package hr.avrbanac.inanis.internals;
 
 import hr.avrbanac.inanis.InanisConfig;
+import hr.avrbanac.inanis.loaders.Loader;
+import hr.avrbanac.inanis.models.RawModel;
+import hr.avrbanac.inanis.models.TestQuadModel;
+import hr.avrbanac.inanis.renderers.Renderer;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.lwjgl.glfw.GLFW;
@@ -28,16 +32,26 @@ public class Engine implements Runnable {
     @Override
     public void run() {
         initDisplay();
-        while(running) {
 
-            update();
+        Loader loader = new Loader();
+        Renderer renderer = new Renderer();
+        renderer.init();
+        RawModel model = loader.loadToVAO(TestQuadModel.QUAD_VERTICES);
+
+        while(running) {
+            renderer.prepare();
+            renderer.render(model);
             render();
+            update();
 
             // escape key pressed or close window button
             if(GLFW.glfwWindowShouldClose(displayManager.getWindowId())) {
                 this.stop();
             }
         }
+
+        loader.cleanUp();
+        displayManager.closeDisplay();
     }
 
     public boolean isRunning() {